requirements.txt
feedparser==6.0.10
yfinance==0.2.18
pandas==2.0.3
smartapi-python==1.5.5
pyotp==2.8.0
websocket-client==1.6.2
psycopg2-binary==2.9.6
pytz==2023.3
logzero
bse
webdriver-manager 
beautifulsoup4
xgboost
scikit-learn